* Step 0: Move the `cudaq::measure_result` class to its own header file   in `qis` directory, remove the macro for library mode and drop the   typedef to bool. * Retain variable names assigned to measurement results * Disallow measure_result as argument to entry-point kernels * Drop the intrinsic funciton fro converting to Boolean vector * Insert discriminate op wherever necessary * Adjust tests * WIP

+++ test/AST-Quake/cast.cpp
@@ -6,7 +6,7 @@
  * the terms of the Apache License 2.0 which accompanies this distribution.    *
  ******************************************************************************/
 
-// RUN: cudaq-quake %s | FileCheck %s
+// RUN: cudaq-quake %s | cudaq-opt | FileCheck %s
 
 #include <cudaq.h>
 
@@ -14,7 +14,7 @@ struct testCast {
   void operator()() __qpu__ {
     cudaq::qubit q0, q1;
     h(q0);
-    double bit = mz(q0);
+    double bit = static_cast<double>(mz(q0));
     // This tests implicit casting from double to bool
     if (bit)
       x(q1);
@@ -29,16 +29,16 @@ struct testCast {
 // CHECK:           %[[VAL_2:.*]] = quake.alloca !quake.ref
 // CHECK:           quake.h %[[VAL_1]] : (!quake.ref) -> ()
 // CHECK:           %[[VAL_3:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
-// CHECK:           %[[VAL_10:.*]] = quake.discriminate %[[VAL_3]] :
-// CHECK:           %[[VAL_4:.*]] = cc.cast unsigned %[[VAL_10]] : (i1) -> f64
-// CHECK:           %[[VAL_5:.*]] = cc.alloca f64
-// CHECK:           cc.store %[[VAL_4]], %[[VAL_5]] : !cc.ptr<f64>
-// CHECK:           %[[VAL_6:.*]] = cc.load %[[VAL_5]] : !cc.ptr<f64>
-// CHECK:           %[[VAL_7:.*]] = arith.cmpf une, %[[VAL_6]], %[[VAL_0]] : f64
-// CHECK:           cc.if(%[[VAL_7]]) {
+// CHECK:           %[[VAL_4:.*]] = quake.discriminate %[[VAL_3]] : (!quake.measure) -> i1
+// CHECK:           %[[VAL_5:.*]] = cc.cast unsigned %[[VAL_4]] : (i1) -> f64
+// CHECK:           %[[VAL_6:.*]] = cc.alloca f64
+// CHECK:           cc.store %[[VAL_5]], %[[VAL_6]] : !cc.ptr<f64>
+// CHECK:           %[[VAL_7:.*]] = cc.load %[[VAL_6]] : !cc.ptr<f64>
+// CHECK:           %[[VAL_8:.*]] = arith.cmpf une, %[[VAL_7]], %[[VAL_0]] : f64
+// CHECK:           cc.if(%[[VAL_8]]) {
 // CHECK:             quake.x %[[VAL_2]] : (!quake.ref) -> ()
 // CHECK:           }
-// CHECK:           %[[VAL_8:.*]] = quake.mz %[[VAL_2]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_9:.*]] = quake.mz %[[VAL_2]] : (!quake.ref) -> !quake.measure
 // CHECK:           return
 // CHECK:         }
 // clang-format on
@@ -58,24 +58,24 @@ struct testCastBoolMeasurement {
 
 // clang-format off
 // CHECK-LABEL:   func.func @__nvqpp__mlirgen__testCastBoolMeasurement() attributes {"cudaq-entrypoint", "cudaq-kernel"} {
-// CHECK:           %[[VAL_1_i32:.*]] = arith.constant 1 : i32
-// CHECK:           %[[VAL_0:.*]] = quake.alloca !quake.ref
+// CHECK:           %[[VAL_0:.*]] = arith.constant 1 : i32
 // CHECK:           %[[VAL_1:.*]] = quake.alloca !quake.ref
-// CHECK:           quake.h %[[VAL_0]] : (!quake.ref) -> ()
-// CHECK:           %[[VAL_m0:.*]] = quake.mz %[[VAL_0]] name {{.*}} : (!quake.ref) -> !quake.measure
-// CHECK:           %[[VAL_2:.*]] = quake.discriminate %[[VAL_m0]] :
-// CHECK:           %[[VAL_3:.*]] = cc.alloca i1
-// CHECK:           cc.store %[[VAL_2]], %[[VAL_3]] : !cc.ptr<i1>
-// CHECK:           %[[VAL_4:.*]] = cc.load %[[VAL_3]] : !cc.ptr<i1>
-// CHECK:           %[[VAL_5:.*]] = cc.cast unsigned %[[VAL_4]] : (i1) -> i32
-// CHECK:           %[[VAL_6:.*]] = cc.alloca i32
-// CHECK:           cc.store %[[VAL_5]], %[[VAL_6]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_7:.*]] = cc.load %[[VAL_6]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_8:.*]] = arith.cmpi eq, %[[VAL_7:.*]], %[[VAL_1_i32]] : i32
-// CHECK:           cc.if(%[[VAL_8]]) {
-// CHECK:             quake.x %[[VAL_1]] : (!quake.ref) -> ()
+// CHECK:           %[[VAL_2:.*]] = quake.alloca !quake.ref
+// CHECK:           quake.h %[[VAL_1]] : (!quake.ref) -> ()
+// CHECK:           %[[VAL_3:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_4:.*]] = quake.discriminate %[[VAL_3]] : (!quake.measure) -> i1
+// CHECK:           %[[VAL_5:.*]] = cc.alloca i1
+// CHECK:           cc.store %[[VAL_4]], %[[VAL_5]] : !cc.ptr<i1>
+// CHECK:           %[[VAL_6:.*]] = cc.load %[[VAL_5]] : !cc.ptr<i1>
+// CHECK:           %[[VAL_7:.*]] = cc.cast unsigned %[[VAL_6]] : (i1) -> i32
+// CHECK:           %[[VAL_8:.*]] = cc.alloca i32
+// CHECK:           cc.store %[[VAL_7]], %[[VAL_8]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_9:.*]] = cc.load %[[VAL_8]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_10:.*]] = arith.cmpi eq, %[[VAL_9]], %[[VAL_0]] : i32
+// CHECK:           cc.if(%[[VAL_10]]) {
+// CHECK:             quake.x %[[VAL_2]] : (!quake.ref) -> ()
 // CHECK:           }
-// CHECK:           %[[VAL_m1:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_11:.*]] = quake.mz %[[VAL_2]] : (!quake.ref) -> !quake.measure
 // CHECK:           return
 // CHECK:         }
 // clang-format on
@@ -94,16 +94,16 @@ struct testUnsignedCastBoolConstTrue {
 
 // clang-format off
 // CHECK-LABEL:   func.func @__nvqpp__mlirgen__testUnsignedCastBoolConstTrue() attributes {"cudaq-entrypoint", "cudaq-kernel"} {
-// CHECK:           %[[VAL_1_i32:.*]] = arith.constant 1 : i32
-// CHECK:           %[[VAL_0:.*]] = quake.alloca !quake.ref
-// CHECK:           %[[VAL_1:.*]] = cc.alloca i32
-// CHECK:           cc.store %[[VAL_1_i32]], %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_2:.*]] = cc.load %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_3:.*]] = arith.cmpi eq, %[[VAL_2:.*]], %[[VAL_1_i32]] : i32
-// CHECK:           cc.if(%[[VAL_3]]) {
-// CHECK:             quake.x %[[VAL_0]] : (!quake.ref) -> ()
+// CHECK:           %[[VAL_0:.*]] = arith.constant 1 : i32
+// CHECK:           %[[VAL_1:.*]] = quake.alloca !quake.ref
+// CHECK:           %[[VAL_2:.*]] = cc.alloca i32
+// CHECK:           cc.store %[[VAL_0]], %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_3:.*]] = cc.load %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_4:.*]] = arith.cmpi eq, %[[VAL_3]], %[[VAL_0]] : i32
+// CHECK:           cc.if(%[[VAL_4]]) {
+// CHECK:             quake.x %[[VAL_1]] : (!quake.ref) -> ()
 // CHECK:           }
-// CHECK:           %[[VAL_m:.*]] = quake.mz %[[VAL_0]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_5:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
 // CHECK:           return
 // CHECK:         }
 // clang-format on
@@ -122,16 +122,16 @@ struct testUnsignedCastBoolConstFalse {
 
 // clang-format off
 // CHECK-LABEL:   func.func @__nvqpp__mlirgen__testUnsignedCastBoolConstFalse() attributes {"cudaq-entrypoint", "cudaq-kernel"} {
-// CHECK:           %[[VAL_0_i32:.*]] = arith.constant 0 : i32
-// CHECK:           %[[VAL_0:.*]] = quake.alloca !quake.ref
-// CHECK:           %[[VAL_1:.*]] = cc.alloca i32
-// CHECK:           cc.store %[[VAL_0_i32]], %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_2:.*]] = cc.load %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_3:.*]] = arith.cmpi eq, %[[VAL_2:.*]], %[[VAL_0_i32]] : i32
-// CHECK:           cc.if(%[[VAL_3]]) {
-// CHECK:             quake.x %[[VAL_0]] : (!quake.ref) -> ()
+// CHECK:           %[[VAL_0:.*]] = arith.constant 0 : i32
+// CHECK:           %[[VAL_1:.*]] = quake.alloca !quake.ref
+// CHECK:           %[[VAL_2:.*]] = cc.alloca i32
+// CHECK:           cc.store %[[VAL_0]], %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_3:.*]] = cc.load %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_4:.*]] = arith.cmpi eq, %[[VAL_3]], %[[VAL_0]] : i32
+// CHECK:           cc.if(%[[VAL_4]]) {
+// CHECK:             quake.x %[[VAL_1]] : (!quake.ref) -> ()
 // CHECK:           }
-// CHECK:           %[[VAL_m:.*]] = quake.mz %[[VAL_0]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_5:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
 // CHECK:           return
 // CHECK:         }
 // clang-format on
@@ -150,16 +150,16 @@ struct testSignedCastBoolConstTrue {
 
 // clang-format off
 // CHECK-LABEL:   func.func @__nvqpp__mlirgen__testSignedCastBoolConstTrue() attributes {"cudaq-entrypoint", "cudaq-kernel"} {
-// CHECK:           %[[VAL_1_i32:.*]] = arith.constant 1 : i32
-// CHECK:           %[[VAL_0:.*]] = quake.alloca !quake.ref
-// CHECK:           %[[VAL_1:.*]] = cc.alloca i32
-// CHECK:           cc.store %[[VAL_1_i32]], %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_2:.*]] = cc.load %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_3:.*]] = arith.cmpi eq, %[[VAL_2:.*]], %[[VAL_1_i32]] : i32
-// CHECK:           cc.if(%[[VAL_3]]) {
-// CHECK:             quake.x %[[VAL_0]] : (!quake.ref) -> ()
+// CHECK:           %[[VAL_0:.*]] = arith.constant 1 : i32
+// CHECK:           %[[VAL_1:.*]] = quake.alloca !quake.ref
+// CHECK:           %[[VAL_2:.*]] = cc.alloca i32
+// CHECK:           cc.store %[[VAL_0]], %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_3:.*]] = cc.load %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_4:.*]] = arith.cmpi eq, %[[VAL_3]], %[[VAL_0]] : i32
+// CHECK:           cc.if(%[[VAL_4]]) {
+// CHECK:             quake.x %[[VAL_1]] : (!quake.ref) -> ()
 // CHECK:           }
-// CHECK:           %[[VAL_m:.*]] = quake.mz %[[VAL_0]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_5:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
 // CHECK:           return
 // CHECK:         }
 // clang-format on
@@ -178,16 +178,16 @@ struct testSignedCastBoolConstFalse {
 
 // clang-format off
 // CHECK-LABEL:   func.func @__nvqpp__mlirgen__testSignedCastBoolConstFalse() attributes {"cudaq-entrypoint", "cudaq-kernel"} {
-// CHECK:           %[[VAL_0_i32:.*]] = arith.constant 0 : i32
-// CHECK:           %[[VAL_0:.*]] = quake.alloca !quake.ref
-// CHECK:           %[[VAL_1:.*]] = cc.alloca i32
-// CHECK:           cc.store %[[VAL_0_i32]], %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_2:.*]] = cc.load %[[VAL_1]] : !cc.ptr<i32>
-// CHECK:           %[[VAL_3:.*]] = arith.cmpi eq, %[[VAL_2:.*]], %[[VAL_0_i32]] : i32
-// CHECK:           cc.if(%[[VAL_3]]) {
-// CHECK:             quake.x %[[VAL_0]] : (!quake.ref) -> ()
+// CHECK:           %[[VAL_0:.*]] = arith.constant 0 : i32
+// CHECK:           %[[VAL_1:.*]] = quake.alloca !quake.ref
+// CHECK:           %[[VAL_2:.*]] = cc.alloca i32
+// CHECK:           cc.store %[[VAL_0]], %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_3:.*]] = cc.load %[[VAL_2]] : !cc.ptr<i32>
+// CHECK:           %[[VAL_4:.*]] = arith.cmpi eq, %[[VAL_3]], %[[VAL_0]] : i32
+// CHECK:           cc.if(%[[VAL_4]]) {
+// CHECK:             quake.x %[[VAL_1]] : (!quake.ref) -> ()
 // CHECK:           }
-// CHECK:           %[[VAL_m:.*]] = quake.mz %[[VAL_0]] : (!quake.ref) -> !quake.measure
+// CHECK:           %[[VAL_5:.*]] = quake.mz %[[VAL_1]] : (!quake.ref) -> !quake.measure
 // CHECK:           return
 // CHECK:         }
 // clang-format on
